+++ docs/metro/js/metro.js
@@ -106,8 +106,8 @@ var isTouch = (('ontouchstart' in window) || (navigator.MaxTouchPoints > 0) || (
 
 var Metro = {
 
-    version: "4.2.39-dev 12/03/2019 14:56",
-    versionFull: "4.2.39-dev 12/03/2019 14:56",
+    version: "@@version",
+    versionFull: "@@version.@@build @@status",
     isTouchable: isTouch,
     fullScreenEnabled: document.fullscreenEnabled,
     sheet: null,
@@ -10055,6 +10055,7 @@ var Draggable = {
             cursor: 'default',
             zIndex: '0'
         };
+        this.dragArea = null;
 
         this._setOptionsFromDOM();
         this._create();
@@ -10090,7 +10091,7 @@ var Draggable = {
 
     _create: function(){
         var that = this, element = this.element, elem = this.elem, o = this.options;
-        var dragArea;
+        var offset = element.offset();
         var position = {
             x: 0,
             y: 0
@@ -10099,15 +10100,25 @@ var Draggable = {
 
         dragElement[0].ondragstart = function(){return false;};
 
+        element.css("position", "absolute");
+
+        if (o.dragArea === 'document' || o.dragArea === 'window') {
+            o.dragArea = "body";
+        }
+
+        this.dragArea = o.dragArea === 'parent' ? element.parent() : $(o.dragArea);
+
+        if (o.dragArea !== 'parent') {
+            element.appendTo(this.dragArea);
+            element.css({
+                top: offset.top,
+                left: offset.left
+            });
+        }
+
         dragElement.on(Metro.events.start, function(e){
 
-            if (o.dragArea === 'document' || o.dragArea === 'window') {
-                o.dragArea = "body";
-            }
-
-            dragArea = o.dragArea === 'parent' ? element.parent() : $(o.dragArea);
-
-            var coord = o.dragArea === "body" ? element.offset() : element.position(),
+            var coord = o.dragArea !== "parent" ? element.offset() : element.position(),
                 shiftX = Utils.pageXY(e).x - coord.left,
                 shiftY = Utils.pageXY(e).y - coord.top;
 
@@ -10118,8 +10129,8 @@ var Draggable = {
                 if (top < 0) top = 0;
                 if (left < 0) left = 0;
 
-                if (top > dragArea.outerHeight() - element.outerHeight()) top = dragArea.outerHeight() - element.outerHeight();
-                if (left > dragArea.outerWidth() - element.outerWidth()) left = dragArea.outerWidth() - element.outerWidth();
+                if (top > that.dragArea.outerHeight() - element.outerHeight()) top = that.dragArea.outerHeight() - element.outerHeight();
+                if (left > that.dragArea.outerWidth() - element.outerWidth()) left = that.dragArea.outerWidth() - element.outerWidth();
 
                 position.y = top;
                 position.x = left;
@@ -10144,9 +10155,7 @@ var Draggable = {
             that.backup.cursor = element.css("cursor");
             that.backup.zIndex = element.css("z-index");
 
-            element.css("position", "absolute").addClass("draggable");
-
-            element.appendTo(dragArea);
+            element.addClass("draggable");
 
             moveElement(e);